resolve bug #23.  Thanks for highlighting this @danikdanik. Keep the issues flowing, when I get moments I always come back to check them out and hopefully resolve them without too much drama :-). Cheers

+++ virtualbox/library_ext/machine.py
@@ -56,8 +56,9 @@ class IMachine(library.IMachine):
             progress.wait_for_completion(-1)
             media = []
         
-        # if delete - let's remove the settings folder too
-        if delete:
+        # if delete - At some virtualbox didn't do a full cleanup of this dir.
+        #             Let's double check it has been cleaned up.
+        if delete and os.path.exists(settings_dir):
             shutil.rmtree(settings_dir)
 
         return media
